[MODINVOSTO-130] - Create batch-voucher API Fix XSD and Spring configs
