+++ ramls/schemas/batch_voucher.xsd
@@ -70,7 +70,7 @@
           <xs:documentation>The currency in which invoice was originally defined</xs:documentation>
         </xs:annotation>
       </xs:element>
-      <xs:element name="invoiceNote" type="xs:string" nillable="true" minOccurs="0">
+      <xs:element name="invoiceNote" type="xs:string" minOccurs="0">
         <xs:annotation>
           <xs:documentation>Free-form notes from the invoice</xs:documentation>
         </xs:annotation>
