+++ src/main/java/org/folio/converters/BatchedVoucherLineModelConverter.java
@@ -1,0 +1,20 @@
+package org.folio.converters;
+
+import org.folio.rest.jaxrs.model.BatchedVoucherLine;
+import org.folio.rest.jaxrs.model.jaxb.BatchedVoucherLineType;
+import org.springframework.core.convert.converter.Converter;
+import java.math.BigDecimal;
+
+public class BatchedVoucherLineModelConverter implements Converter<BatchedVoucherLine, BatchedVoucherLineType> {
+  @Override
+  public BatchedVoucherLineType convert(BatchedVoucherLine batchedVoucherLine) {
+    BatchedVoucherLineType batchedVoucherLineType = new BatchedVoucherLineType();
+    batchedVoucherLineType.setAmount(BigDecimal.valueOf(batchedVoucherLine.getAmount()));
+    batchedVoucherLineType.setExternalAccountNumber(batchedVoucherLine.getExternalAccountNumber());
+
+    BatchedVoucherLineType.FundCodes fundCodes = new BatchedVoucherLineType.FundCodes();
+    fundCodes.withFundCode(batchedVoucherLine.getFundCodes());
+    batchedVoucherLineType.withFundCodes(fundCodes);
+    return batchedVoucherLineType;
+  }
+}
