+++ src/main/java/org/folio/jaxb/JAXBUtil.java
@@ -1,14 +1,21 @@
 package org.folio.jaxb;
 
+import javax.xml.datatype.DatatypeConfigurationException;
+import javax.xml.datatype.DatatypeFactory;
+import javax.xml.datatype.XMLGregorianCalendar;
 import javax.xml.transform.stream.StreamSource;
+import java.time.Instant;
+import java.time.format.DateTimeFormatter;
 import java.util.Arrays;
 import java.util.Collections;
+import java.util.Date;
 import java.util.List;
 import java.util.Objects;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
 public class JAXBUtil {
+  private static final DateTimeFormatter fromFormatter = DateTimeFormatter.ofPattern("yyyy-MM-dd'T'HH:mm:ss.SSSZ");
 
   public static List<StreamSource> xsdSchemasAsStreamResources(final String schemasDirectory, final String[] fileNames){
     if (Objects.nonNull(schemasDirectory) && Objects.nonNull(fileNames)){
@@ -31,6 +38,38 @@ public class JAXBUtil {
                    .collect(Collectors.toList());
     }
     return Collections.emptyList();
+  }
+
+  /**
+   *
+   * @param dateTime  "2019-12-06T00:00:00.000+0000"
+   *
+   * @return "2019-12-06T00:01:04Z"
+   */
+  public static XMLGregorianCalendar convertDateTime(String dateTime) {
+    Instant instant = Instant.from(fromFormatter.parse(dateTime));
+    return getXmlGregorianCalendar(instant);
+  }
+
+  /**
+   *
+   * @param dateTime  "2019-12-06T00:00:00.000+0000"
+   *
+   * @return "2019-12-06T00:01:04Z"
+   */
+  public static XMLGregorianCalendar convertOldJavaDate(Date dateTime) {
+    Instant instant = dateTime.toInstant();
+    return getXmlGregorianCalendar(instant);
+  }
+
+  private static XMLGregorianCalendar getXmlGregorianCalendar(Instant instant) {
+    XMLGregorianCalendar result;
+    try {
+      result = DatatypeFactory.newInstance().newXMLGregorianCalendar(instant.toString());
+    } catch (DatatypeConfigurationException e) {
+      throw new IllegalArgumentException(e);
+    }
+    return result;
   }
 
   private static Class<?> loadClass(String className) {
